Output/primary/tests/test_arraylist.rs
use primary::arraylist::*;

pub struct ArrayList {
    // Define the fields of ArrayList here
}

impl ArrayList {
    pub fn new(capacity: usize) -> Self {
        // Implement the constructor here
    }

    pub fn append(&mut self, item: &i32) {
        // Implement the append method here
    }
}

pub fn generate_arraylist() -> ArrayList {
    let mut arraylist = ArrayList::new(0);

    for _ in 0..4 {
        arraylist.append(&variable1);
        arraylist.append(&variable2);
        arraylist.append(&variable3);
        arraylist.append(&variable4);
    }

    arraylist
}

pub static mut VARIABLE1: i32 = 0;
pub static mut VARIABLE2: i32 = 0;
pub static mut VARIABLE3: i32 = 0;
pub static mut VARIABLE4: i32 = 0;
#[test]
pub fn test_arraylist_new_free() {
    // Use a default size when given zero
    let mut arraylist = arraylist_new(0);
    assert!(arraylist.is_some());
    arraylist_free(&mut arraylist);

    // Normal allocated
    arraylist = arraylist_new(10);
    assert!(arraylist.is_some());
    arraylist_free(&mut arraylist);

    // Freeing a null arraylist works
    arraylist_free(&mut None);

    // Test low memory scenarios (failed malloc)
    alloc_test_set_limit(0);
    arraylist = arraylist_new(0);
    assert!(arraylist.is_none());

    alloc_test_set_limit(1);
    arraylist = arraylist_new(100);
    assert!(arraylist.is_none());
}